Change trigger and death logic.

+++ Assets/CodeBase/Logic/SpeakerText.cs
@@ -7,11 +7,11 @@ public class SpeakerText
 {
     [SerializeField] private TextMeshProUGUI _text;
     [SerializeField] private Vector3 _targetTextSize;
-    [SerializeField] private int _timeTextShow;
-    [SerializeField] private int _durationTextShow;
+    [SerializeField] private float _timeTextShow;
+    [SerializeField] private float _durationTextShow;
 
     public TextMeshProUGUI Text => _text;
     public Vector3 TargetTextSize => _targetTextSize;
-    public int TimeTextShow => _timeTextShow;
-    public int DurationTextShow => _durationTextShow;
+    public float TimeTextShow => _timeTextShow;
+    public float DurationTextShow => _durationTextShow;
 }
